+++ gneiss-mqtt-aws/CHANGELOG.md
@@ -16,3 +16,4 @@ This document is currently hand-written and non-authoritative.
 * Features
 * * Rustls support feature-gated
 * * Native-tls support added and feature-gated
+* * Websocket support feature-gated